[refactor] #51 : find hooks 분리
